services/auth/requirements.txt
Django>=4.2,<5
django-ninja>=1.0
djangorestframework-simplejwt>=5.3
PyJWT>=2.8
gunicorn>=21.0
